refactor: 기능별 패키지 구조로 코드 리팩토링 - 패키지명은 각각 teamup은 팀원 모집, findproject는 프로젝트 찾기에 해당
