+++ src/main/java/ssu/cromi/teamit/controller/findproject/ProjectDetailController.java
@@ -1,11 +1,11 @@
-package ssu.cromi.teamit.controller;
+package ssu.cromi.teamit.controller.findproject;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import ssu.cromi.teamit.DTO.ProjectDetailResponseDto;
-import ssu.cromi.teamit.service.ProjectDetailService;
+import ssu.cromi.teamit.service.findproject.ProjectDetailService;
 
 @RestController
 @RequestMapping("/v1/projects")
