+++ src/main/java/ssu/cromi/teamit/repository/findproject/ProjectApplicationRepository.java
@@ -1,7 +1,7 @@
-package ssu.cromi.teamit.repository;
+package ssu.cromi.teamit.repository.findproject;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import ssu.cromi.teamit.entity.ProjectApplication;
+import ssu.cromi.teamit.entity.findproject.ProjectApplication;
 
 public interface ProjectApplicationRepository extends JpaRepository<ProjectApplication, Long> {
 }
