+++ src/main/java/ssu/cromi/teamit/repository/teamup/ProjectMemberRepository.java
@@ -1,7 +1,7 @@
-package ssu.cromi.teamit.repository;
+package ssu.cromi.teamit.repository.teamup;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import ssu.cromi.teamit.entity.ProjectMember;
+import ssu.cromi.teamit.entity.teamup.ProjectMember;
 
 public interface ProjectMemberRepository extends JpaRepository<ProjectMember, Long> {
 }
