+++ src/main/java/ssu/cromi/teamit/service/findproject/ProjectApplicationServiceImpl.java
@@ -1,4 +1,4 @@
-package ssu.cromi.teamit.service;
+package ssu.cromi.teamit.service.findproject;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -6,10 +6,10 @@ import org.springframework.transaction.annotation.Transactional;
 
 import ssu.cromi.teamit.exception.ProjectNotFoundException;
 import ssu.cromi.teamit.DTO.ProjectApplicationRequestDto;
-import ssu.cromi.teamit.entity.Project;
-import ssu.cromi.teamit.entity.ProjectApplication;
-import ssu.cromi.teamit.repository.ProjectApplicationRepository;
-import ssu.cromi.teamit.repository.ProjectRepository;
+import ssu.cromi.teamit.entity.teamup.Project;
+import ssu.cromi.teamit.entity.findproject.ProjectApplication;
+import ssu.cromi.teamit.repository.findproject.ProjectApplicationRepository;
+import ssu.cromi.teamit.repository.teamup.ProjectRepository;
 
 import java.time.LocalDateTime;
