+++ src/main/java/ssu/cromi/teamit/service/findproject/ProjectDetailService.java
@@ -1,4 +1,4 @@
-package ssu.cromi.teamit.service;
+package ssu.cromi.teamit.service.findproject;
 
 import ssu.cromi.teamit.DTO.ProjectDetailResponseDto;
